+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxInfo.hs
@@ -118,6 +118,7 @@ import Cardano.Ledger.Plutus.Language (
   SLanguage (..),
   decodePlutusRunnable,
  )
+import qualified Cardano.Ledger.Plutus.PolicyID.Translation as PlutusPolicyID
 import Cardano.Ledger.Plutus.ToPlutusData (ToPlutusData (..))
 import Cardano.Ledger.Plutus.TxInfo (
   slotToPOSIXTime,
@@ -133,6 +134,8 @@ import Cardano.Ledger.Plutus.TxInfo (
   transScriptHash,
  )
 import qualified Cardano.Ledger.Plutus.TxInfo as TxInfo
+import qualified Cardano.Ledger.Plutus.Value.Translation.V1V2 as PlutusV1V2
+import qualified Cardano.Ledger.Plutus.Value.Translation.V3V4 as PlutusV3V4
 import Cardano.Ledger.TxIn (TxId (..), TxIn (..))
 import Cardano.Slotting.EpochInfo (EpochInfo)
 import Cardano.Slotting.Time (SystemStart)
@@ -151,7 +154,6 @@ import Lens.Micro ((^.))
 import qualified PlutusLedgerApi.V1 as PV1
 import qualified PlutusLedgerApi.V2 as PV2
 import qualified PlutusLedgerApi.V3 as PV3
-import qualified PlutusLedgerApi.V3.MintValue as PV3
 
 instance EraPlutusContext ConwayEra where
   type ContextError ConwayEra = ConwayContextError ConwayEra
@@ -438,7 +440,7 @@ instance EraPlutusTxInfo 'PlutusV1 ConwayEra where
             { PV1.txInfoInputs = inputs
             , PV1.txInfoOutputs = outputs
             , PV1.txInfoFee = transCoinToValue (txBody ^. feeTxBodyL)
-            , PV1.txInfoMint = Alonzo.transMintValue (txBody ^. mintTxBodyL)
+            , PV1.txInfoMint = PlutusV1V2.fromLedgerForging (txBody ^. forgingTxBodyL)
             , PV1.txInfoDCert = txCerts
             , PV1.txInfoWdrl = Alonzo.transTxBodyWithdrawals txBody
             , PV1.txInfoValidRange = timeRange
@@ -484,7 +486,7 @@ instance EraPlutusTxInfo 'PlutusV2 ConwayEra where
             , PV2.txInfoOutputs = outputs
             , PV2.txInfoReferenceInputs = refInputs
             , PV2.txInfoFee = transCoinToValue (txBody ^. feeTxBodyL)
-            , PV2.txInfoMint = Alonzo.transMintValue (txBody ^. mintTxBodyL)
+            , PV2.txInfoMint = PlutusV1V2.fromLedgerForging (txBody ^. forgingTxBodyL)
             , PV2.txInfoDCert = txCerts
             , PV2.txInfoWdrl = PV2.unsafeFromList $ Alonzo.transTxBodyWithdrawals txBody
             , PV2.txInfoValidRange = timeRange
@@ -533,7 +535,7 @@ instance EraPlutusTxInfo 'PlutusV3 ConwayEra where
               , PV3.txInfoOutputs = outputs
               , PV3.txInfoReferenceInputs = refInputsInfo
               , PV3.txInfoFee = transCoinToLovelace (txBody ^. feeTxBodyL)
-              , PV3.txInfoMint = transMintValue (txBody ^. mintTxBodyL)
+              , PV3.txInfoMint = PlutusV3V4.fromLedgerForging (txBody ^. forgingTxBodyL)
               , PV3.txInfoTxCerts = txCerts
               , PV3.txInfoWdrl = transTxBodyWithdrawals txBody
               , PV3.txInfoValidRange = timeRange
@@ -570,8 +572,10 @@ transTxBodyId txBody = PV3.TxId (transSafeHash (hashAnnotated @_ @EraIndependent
 transTxIn :: TxIn -> PV3.TxOutRef
 transTxIn (TxIn txid txIx) = PV3.TxOutRef (transTxId txid) (toInteger (txIxToInt txIx))
 
+-- | Compatibility entry point for the raw Ledger mint representation.
+{-# DEPRECATED transMintValue "Use `Cardano.Ledger.Plutus.Value.Translation.V3V4.fromLedgerForging`" #-}
 transMintValue :: MultiAsset -> PV3.MintValue
-transMintValue = PV3.UnsafeMintValue . PV1.getValue . Alonzo.transMultiAsset
+transMintValue = PlutusV3V4.fromLedgerForging . Forging
 
 -- | Translate all `Withdrawal`s from within a `TxBody`
 transTxBodyWithdrawals :: EraTxBody era => TxBody l era -> PV3.Map PV3.Credential PV3.Lovelace
@@ -664,7 +668,7 @@ transPlutusPurposeV3 ::
   Either (ContextError era) PV3.ScriptPurpose
 transPlutusPurposeV3 proxy pv _ = \case
   SpendingPurpose (AsIxItem _ txIn) -> pure $ PV3.Spending (transTxIn txIn)
-  MintingPurpose (AsIxItem _ policyId) -> pure $ PV3.Minting (Alonzo.transPolicyID policyId)
+  MintingPurpose (AsIxItem _ policyId) -> pure $ PV3.Minting (PlutusPolicyID.fromLedgerPolicyID policyId)
   CertifyingPurpose (AsIxItem ix txCert) ->
     PV3.Certifying (toInteger ix) <$> toPlutusTxCert proxy pv txCert
   WithdrawingPurpose (AsIxItem _ accountAddress) -> pure $ PV3.Rewarding (transAccountAddress accountAddress)